#243 util file for listing helper functions
